reactor: separate EditForm from EditCourse, change type status and change id to course_id
