+++ frontend/src/types/course.d.ts
@@ -8,7 +8,7 @@ export type Category = {
 export type NewCourse = {
     title: string;
     categories: number[];
-    status: number;
+    status: boolean;
     summary: string;
     description: string;
     thumbnail: string;
@@ -71,7 +71,7 @@ export type ChangeThumbnail = {
 };
 
 export type CourseChangeInformation = {
-    id: number | undefined;
+    course_id: number | undefined;
     title: string;
     summary: string;
     status: boolean;
